+++ src/types/__tests__/effector/sample/sample.test.ts
@@ -11,6 +11,19 @@ import {
   EventCallable,
 } from 'effector'
 
+const consoleError = console.error
+
+beforeAll(() => {
+  console.error = (message, ...args) => {
+    if (String(message).includes('guard')) return
+    consoleError(message, ...args)
+  }
+})
+
+afterAll(() => {
+  console.error = consoleError
+})
+
 const typecheck = '{global}'
 
 describe('explicit generics', () => {